+++ finagle-http/src/test/scala/com/twitter/finagle/http/H2CToTlsSnoopingEndToEndTest.scala
@@ -2,7 +2,7 @@ package com.twitter.finagle.http
 
 import com.twitter.finagle
 import com.twitter.finagle.http.ssl.HttpSslTestComponents
-import com.twitter.finagle.ssl.OpportunisticTls
+import com.twitter.finagle.ssl.{OpportunisticTls, SnoopingLevelInterpreter}
 import com.twitter.finagle.transport.Transport
 
 // Test of a H2C cleartext client to the TLS snooping H2 compatible server.
@@ -13,6 +13,7 @@ class H2CToTlsSnoopingEndToEndTest extends H2CEndToEndTest {
   override def serverImpl(): finagle.Http.Server =
     super
       .serverImpl()
+      .configured(SnoopingLevelInterpreter.EnabledForNonNegotiatingProtocols)
       .configured(Transport.ServerSsl(Some(HttpSslTestComponents.unauthenticatedServerConfig)))
       .configured(OpportunisticTls.Param(OpportunisticTls.Desired))
 }